// examples/api-prisma/src/generated/nestjs-graphql-dtos/task/task-scalar-where.input.ts
import { Field, InputType } from '@nestjs/graphql';

import { EnumTaskStatusFilter } from '../prisma/enum-task-status-filter.input';
import { IntFilter } from '../prisma/int-filter.input';
import { StringFilter } from '../prisma/string-filter.input';
import { StringNullableFilter } from '../prisma/string-nullable-filter.input';

@InputType()
export class TaskScalarWhereInput {
  @Field(() => [TaskScalarWhereInput], { nullable: true })
  AND?: Array<TaskScalarWhereInput>;

  @Field(() => [TaskScalarWhereInput], { nullable: true })
  OR?: Array<TaskScalarWhereInput>;

  @Field(() => [TaskScalarWhereInput], { nullable: true })
  NOT?: Array<TaskScalarWhereInput>;

  @Field(() => IntFilter, { nullable: true })
  id?: IntFilter;

  @Field(() => StringFilter, { nullable: true })
  title?: StringFilter;

  @Field(() => StringNullableFilter, { nullable: true })
  description?: StringNullableFilter;

  @Field(() => EnumTaskStatusFilter, { nullable: true })
  status?: EnumTaskStatusFilter;

  @Field(() => IntFilter, { nullable: true })
  authorId?: IntFilter;
}
